#2 Added code for migrate data for specific month only
